refactor(kio): move Backoff out of moq-net's public surface
A backoff schedule is a utility, not part of the wire layer's published
API. It now lives in `kio::time` next to `Deadline`, where the crates
that need it (moq-native, moq-relay, moq-hls, moq-rtmp, moq-audio)
already reach it. `moq_net::retry` is gone.

`status_retryable` moves with its callers rather than being shared:
moq-native and moq-hls each keep a private copy of the five-status list,
so nothing about retry policy is exported from either.

Three review findings folded in:

- The budget now runs from construction or the last `reset` rather than
  from the first delay, so it covers the attempts as well as the waits
  between them. A reconnect whose every dial hung until the connect
  timeout could previously outlive its budget by that much, which also
  put it past the linger window `Client::consume` sizes from it.
- When several HLS renditions fail in one pass, the step keeps the
  failure another pass could clear rather than whichever came last, so a
  single permanently-dead variant doesn't end an import the others could
  still serve. Keeping the last one made the outcome depend on rendition
  order.
- `doc/bin/relay/cluster.md` still promised that a rejected cluster peer
  is given up on, which stopped being true when the classification came
  out. It documents the redial-forever behavior now.
