+++ rs/CLAUDE.md
@@ -109,7 +109,7 @@ Negotiation: `version::NEGOTIATED` lists SETUP-negotiated versions in preference
 
 ## Rust conventions
 
-- **Retries go through `moq_net::retry`** (root Retries explains the why). `retry::Backoff` is the schedule (capped exponential, equal jitter, optional give-up budget): `sleep().await` in an async loop, `delay()` when the caller owns the waiting (a blocking thread, a `select!` arm), `reset()` after an outcome worth trusting. `retry::Config` is `#[non_exhaustive]`, so build it with `default()` + field set. Never hand-roll a `tokio::time::sleep(FIXED)` in a failure arm, and don't add an `is_retryable()` to an error type: the budget is what stops a loop. The one thing worth reading off a failure is a status a peer actually sent, via `retry::status_retryable` and the `status()` accessors on `moq_native::Error` / `moq_hls::Error`.
+- **Retries go through `kio::time::Backoff`** (root Retries explains the why). It's the schedule (capped exponential, equal jitter, optional give-up budget): `sleep().await` in an async loop, `delay()` when the caller owns the waiting (a blocking thread, a `select!` arm), `reset()` after an outcome worth trusting. `kio::time::Config` is `#[non_exhaustive]`, so build it with `default()` + field set. It lives in kio rather than moq-net because a backoff schedule is a utility, not part of the wire layer's surface. Never hand-roll a `tokio::time::sleep(FIXED)` in a failure arm, and don't add an `is_retryable()` to an error type: the budget is what stops a loop. The one thing worth reading off a failure is a status a peer actually sent, via the `status()` accessors on `moq_native::Error` / `moq_hls::Error`.
 - **Prefer `kio` over tokio sync primitives**: reach for `kio::Producer`/`Consumer` (and the `poll_*` plumbing) instead of `tokio::sync` channels or `watch`. A `tokio::sync::watch` (or a channel) carrying a single value is a code smell. `kio` ties into the runtime-free `poll_*` model and avoids a hard runtime dependency.
 - **Errors**: `thiserror` with `#[from]` for libraries, `anyhow` (with `.context("...")`, not `.map_err(|_| anyhow!())`) for binaries. Always `#[non_exhaustive]` on public error enums (e.g. `moq-net/src/error.rs`, `moq-ffi/src/error.rs`, `moq-loc/src/lib.rs`). Use `#[error(transparent)]` + `#[from]` for wrapped foreign errors (see `moq-token/src/error.rs`).
 - **Config + TOML merge**: any `#[arg]` field on a TOML-loadable config must be `Option<T>`, never a bare `bool`/`String`/etc. The TOML->CLI merge re-applies clap defaults and silently clobbers TOML values for bare fields. See `moq-relay/src/config.rs` and its regression tests (`cli_does_not_clobber_toml_*`); add such a test for any new flag.
