+++ rs/moq-native/src/error.rs
@@ -1,5 +1,14 @@
 use std::sync::Arc;
 
+/// Whether an HTTP response status means "ask again later".
+///
+/// A response that arrived is the server's answer, and only this narrow set invites another
+/// attempt: request timeout, rate limit, and the gateway/overload statuses. Every other status,
+/// `404` and `403` included, is settled.
+pub(crate) fn status_retryable(status: u16) -> bool {
+	matches!(status, 408 | 429 | 502 | 503 | 504)
+}
+
 /// Errors produced while configuring or establishing native MoQ connections.
 ///
 /// Backend-specific failures live in per-backend error types ([`crate::tls::Error`],
@@ -150,7 +159,7 @@ impl Error {
 	///
 	/// `None` covers everything else: a dial that never got a response, a QUIC handshake that
 	/// failed, a URL we couldn't parse. Only a status the peer actually sent shows up here, and
-	/// [`moq_net::retry::status_retryable`] is what decides whether it invites another attempt. This
+	/// [`status_retryable`] is what decides whether it invites another attempt. This
 	/// deliberately does not try to say whether some *other* kind of failure is worth retrying;
 	/// that's a guess, and the caller's backoff budget bounds it instead.
 	pub fn status(&self) -> Option<u16> {
@@ -160,11 +169,7 @@ impl Error {
 			// over QUIC alongside a dead WebSocket is still just a failed dial.
 			#[cfg(feature = "websocket")]
 			Self::TransportRace { quic, websocket } => match (quic.status(), websocket.status()) {
-				(Some(quic), Some(websocket))
-					if !moq_net::retry::status_retryable(quic) && !moq_net::retry::status_retryable(websocket) =>
-				{
-					Some(quic)
-				}
+				(Some(quic), Some(websocket)) if !status_retryable(quic) && !status_retryable(websocket) => Some(quic),
 				_ => None,
 			},
