+++ rs/moq-net/src/lib.rs
@@ -84,7 +84,6 @@ mod setup;
 mod util;
 mod version;
 
-pub mod retry;
 pub mod stats;
 
 pub use client::*;
