+++ rs/moq-relay/Cargo.toml
@@ -50,6 +50,7 @@ http-body = "1"
 http-cache-reqwest = { version = "1.0.0-alpha.6", features = ["manager-moka", "url-standard"], default-features = false }
 humantime = "2.3"
 humantime-serde = "1.1"
+kio = { workspace = true, features = ["time"] }
 jsonwebtoken = "11"
 moq-native = { workspace = true, default-features = false, features = ["aws-lc-rs", "watch", "tcp"] }
 moq-net = { workspace = true }
